Add files via upload
EECS355 Final Project: FPGA-based Tank Duel Game
Objective:
In this project, you will design a game called “tank duel” that could be displayed on the VGA monitor and controlled by two players by one PS/2 keyboard.
Requirements:
1. Draw two tanks facing each other on VGA screen with different colors (e.g.
red and blue). One tank should appear on the top of the screen and the other should appear on the bottom. The tanks can be drawn as simple rectangles or any other complicated shapes that you choose. Draw whatever background you like for the battle field.
2. Both tanks should be continuously moving back and forth on a horizontal line on the screen. That means when they reach one end of the screen they should automatically move in the opposite directions without any external input,
3. Both tanks should have 3 different moving speeds that are adjusted by keyboard inputs by the player.
4. Both tanks shoot one bullet at a time; if it hits the opponent’s tank, it scores one point. A tank can shoot the second bullet only when the previous bullet goes out of the screen or hits the opponent.
5. The score of both players should be displayed on the LEDs. Whoever first scores 3 points wins.
6. When one player reaches three points:
a. Only the winner’s tank should be displayed on the screen.
b. A message declaring the winner should print out on the LCD screen.
7. There should be a global reset signal to reset the game;